flow: Make flow_extract()'s caller responsible for metadata.
We are getting to have a lot of metadata in struct flow.  Until now, we
have tried to pass all of it as parameters to flow_extract().  Now that
the data and metadata members of struct flow are clearly separated, it
seems more reasonable to just make the caller initialize them itself.

Signed-off-by: Ben Pfaff <blp@nicira.com>